[Garima] refactor to correct spelling and naming convention
