+++ assets/svg/TicketConductor.tsx
