+++ assets/svg/TicketConfirm.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from "react";
 import Svg, { Circle, Path, SvgProps } from "react-native-svg";
 
-const TciketConfirm = (props?:SvgProps ): ReactElement => {
+const TicketConfirm = (props?:SvgProps ): ReactElement => {
   return (
 <Svg width="92" height="91" viewBox="0 0 92 91" fill="none" { ...props }>
 <Path fill-rule="evenodd" clip-rule="evenodd" d="M8.56362 11.6971C8.3408 13.3544 8.33319 15.6037 8.33319 19.0834L8.33319 64.2784L15.4999 61.8895L26.2499 65.4728L36.9999 61.8895L44.1665 64.2784L44.1665 15.5C44.1665 12.8893 44.8645 10.4416 46.0841 8.33337L19.0832 8.33337C15.6035 8.33337 13.3542 8.34098 11.6969 8.5638C10.1426 8.77276 9.6555 9.11011 9.38272 9.3829C9.10994 9.65569 8.77258 10.1428 8.56362 11.6971ZM58.4999 8.33337C54.5418 8.33337 51.3332 11.542 51.3332 15.5L51.3332 22.6667L65.6665 22.6667L65.6665 15.5C65.6665 11.542 62.4579 8.33337 58.4999 8.33337ZM51.3332 29.8333L66.1784 29.8333C66.211 29.8333 66.2441 29.8334 66.2776 29.8334C66.7865 29.8336 67.3939 29.8339 67.9259 29.7823C68.5481 29.7219 69.433 29.5655 70.3179 29.0095C71.0016 28.5799 71.5797 28.0017 72.0093 27.318C72.5653 26.4331 72.7218 25.5483 72.7821 24.9261C72.8337 24.394 72.8335 23.7867 72.8332 23.2778C72.8332 23.2443 72.8332 23.2112 72.8332 23.1786L72.8332 15.5C72.8332 7.58395 66.4159 1.16669 58.4999 1.1667L18.8474 1.16669C15.6708 1.1665 12.9337 1.16636 10.7419 1.46104C8.393 1.77684 6.14141 2.48901 4.31512 4.3153C2.48883 6.14159 1.77666 8.39318 1.46086 10.7421C1.16618 12.9339 1.16633 15.6709 1.16651 18.8475L1.16652 74.2216L15.4999 69.4438L26.2499 73.0272L36.9999 69.4438L51.3332 74.2216L51.3332 29.8333Z" fill="#8BB19C" stroke="#F7F7F7"/>
@@ -15,4 +15,4 @@ const TciketConfirm = (props?:SvgProps ): ReactElement => {
   )
 }
 
-export default TciketConfirm;
+export default TicketConfirm;
