+++ assets/theme/colors.js
@@ -1,5 +1,5 @@
 const Dim_Black ="#424242";
-const GreyBlack = "#3E4059";
+const Grey_Black = "#3E4059";
 const White = "#ffff";
 const Beige = "#F3F1E0";
 const Black = "#000000";
@@ -9,7 +9,7 @@ const Disabled_Button = "#AAABB5";
 const Transparent = "transparent";
 export const colors = {
   Dim_Black,
-  GreyBlack,
+  Grey_Black,
   White,
   Beige,
   Black,
